서버/클라이언트 코드 분리 테스트
