+++ packages/i18n/client.d.ts
@@ -1,0 +1,1 @@
+export * from './lib/app-dir/client'
