+++ packages/i18n/client.js
@@ -1,0 +1,1 @@
+module.exports = require('./lib/app-dir/client')
