+++ packages/i18n/src/app-dir/constants.ts
@@ -1,0 +1,3 @@
+export const FALLBACK_LANGUAGE = 'ko'
+export const LANGUAGES = [FALLBACK_LANGUAGE, 'en', 'ja', 'zh']
+export const DEFAULT_NAMESPACE = 'local'
