+++ packages/i18n/src/app-dir/types.ts
@@ -1,0 +1,1 @@
+export type Language = 'ko' | 'en' | 'zh' | 'ja'
