+++ packages/i18n/src/index.ts
@@ -3,5 +3,3 @@ import i18nConfig from './next-i18next.config'
 export { default as I18nextTripleWebAssetsBackend } from './backend'
 export { I18nCommonWebKeys } from './i18n-common-web-keys'
 export { i18nConfig }
-
-export * from './app-dir-i18n'
